updated paths to diplomportal and added apiController->service structure to auth
