+++ ReactApp/.env.production
@@ -1,4 +1,4 @@
 // Environment variables for React app PROD
 
-REACT_APP_BASE_URL=https://kvissen.herokuapp.com
+REACT_APP_BASE_URL=http://kvissen.devops.diplomportal.dk:8080
 REACT_APP_LOGIN_URL=/api/auth/login
